requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.8
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
async-timeout==5.0.1
attrs==25.3.0
backoff==2.2.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dataclasses==0.6
datasets==3.6.0
dill==0.3.8
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.6.2
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
gql==3.5.3
graphql-core==3.2.6
h11==0.16.0
halo==0.0.31
hf-xet==1.1.3
huggingface-hub==0.32.4
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
log-symbols==0.0.14
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pluggy==1.6.0
propcache==0.3.1
pyarrow==20.0.0
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pytest==8.4.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
spinners==0.0.24
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
termcolor==3.1.0
tomli==2.2.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transform==1.2.0
transform-tools==1.0.5
triton==3.3.0
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
update-checker==0.18.0
urllib3==2.4.0
uvicorn==0.34.3
validators==0.35.0
xxhash==3.5.0
yarl==1.20.0
